+++ json_serializable/test/kitchen_sink/strict_keys_object.g.dart
@@ -11,7 +11,9 @@ part of 'strict_keys_object.dart';
 // **************************************************************************
 
 StrictKeysObject _$StrictKeysObjectFromJson(Map json) {
-  $checkKeys(json, allowedKeys: const ['value', 'custom_field']);
+  $checkKeys(json,
+      allowedKeys: const ['value', 'custom_field'],
+      requiredKeys: const ['value', 'custom_field']);
   return new StrictKeysObject(
       json['value'] as int, json['custom_field'] as String);
 }